Add 'format' cmake target
To make the cmake project more self-contained, add 'format' target
for executing clang-format.
The Format.py script now wraps this usage, just like with other usages.

The script adds possibility to format a subset of files. This can as well
be achieved using cmake, but it's gonna be uglier and it's not necessary.